Refactor index tests.
