+++ src/include/storage/index/index.h
@@ -28,7 +28,7 @@ class Index {
   friend class GenericKey<64>;
   friend class GenericKey<128>;
   friend class GenericKey<256>;
-  friend class BwTreeIndexTests;
+  friend class BwTreeKeyTests;
 
   const catalog::index_oid_t oid_;
   const ConstraintType constraint_type_;
