+++ src/include/storage/index/index_metadata.h
@@ -92,9 +92,9 @@ class IndexMetadata {
 
  private:
   DISALLOW_COPY(IndexMetadata);
-  FRIEND_TEST(BwTreeIndexTests, IndexMetadataCompactIntsKeyTest);
-  FRIEND_TEST(BwTreeIndexTests, IndexMetadataGenericKeyNoMustInlineVarlenTest);
-  FRIEND_TEST(BwTreeIndexTests, IndexMetadataGenericKeyMustInlineVarlenTest);
+  FRIEND_TEST(BwTreeKeyTests, IndexMetadataCompactIntsKeyTest);
+  FRIEND_TEST(BwTreeKeyTests, IndexMetadataGenericKeyNoMustInlineVarlenTest);
+  FRIEND_TEST(BwTreeKeyTests, IndexMetadataGenericKeyMustInlineVarlenTest);
 
   std::vector<IndexKeyColumn> key_schema_;                                      // for GenericKey
   std::vector<uint8_t> attr_sizes_;                                             // for CompactIntsKey
